+++ src/Conductor/Compilation/ImageCompiler.php
@@ -78,7 +78,8 @@ class ImageCompiler implements CompilerInterface
         array &$definitions,
         CompiledDeploymentInterface $compiledDeployment,
         JobWorkspaceInterface $workspace,
-        JobUnitInterface $job
+        JobUnitInterface $job,
+        ?string $storageIdentifier = null
     ): CompilerInterface {
         $imagesConfigs = $this->mergeConfigurations($definitions, $this->imagesLibrary);