+++ src/Conductor/Compilation/IngressCompiler.php
@@ -52,7 +52,8 @@ class IngressCompiler implements CompilerInterface
         array &$definitions,
         CompiledDeploymentInterface $compiledDeployment,
         JobWorkspaceInterface $workspace,
-        JobUnitInterface $job
+        JobUnitInterface $job,
+        ?string $storageIdentifier = null
     ): CompilerInterface {
         foreach ($definitions as $name => &$config) {
             $paths = [];