+++ src/Conductor/Compilation/ServiceCompiler.php
@@ -48,7 +48,8 @@ class ServiceCompiler implements CompilerInterface
         array &$definitions,
         CompiledDeploymentInterface $compiledDeployment,
         JobWorkspaceInterface $workspace,
-        JobUnitInterface $job
+        JobUnitInterface $job,
+        ?string $storageIdentifier = null
     ): CompilerInterface {
         foreach ($definitions as $name => &$config) {
             $ports = [];