+++ src/Conductor/Compilation/VolumeCompiler.php
@@ -47,7 +47,8 @@ class VolumeCompiler implements CompilerInterface
         array &$definitions,
         CompiledDeploymentInterface $compiledDeployment,
         JobWorkspaceInterface $workspace,
-        JobUnitInterface $job
+        JobUnitInterface $job,
+        ?string $storageIdentifier = null
     ): CompilerInterface {
         foreach ($definitions as $name => &$config) {
             $volumeName = ($config['name'] ?? $name);