+++ src/Contracts/Conductor/CompilerInterface.php
@@ -41,6 +41,7 @@ interface CompilerInterface
         array &$definitions,
         CompiledDeploymentInterface $compiledDeployment,
         JobWorkspaceInterface $workspace,
-        JobUnitInterface $job
+        JobUnitInterface $job,
+        ?string $storageIdentifier = null
     ): CompilerInterface;
 }